requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-bigquery==1.17.0
google-cloud-core==1.0.2
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
iso8601==0.1.12
isort==4.3.21
lazy-object-proxy==1.4.1
lxml==4.3.4
mccabe==0.6.1
parsel==1.5.1
pkg-resources==0.0.0
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pytz==2019.1
queuelib==1.5.0
requests==2.22.0
rsa==4.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
Twisted==19.2.1
typed-ast==1.4.0
urllib3==1.25.3
w3lib==1.20.0
wrapt==1.11.2
zope.interface==4.6.0
